requirements.txt
# Core dependencies
notion-client>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
rich>=13.5.0
pydantic>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0

# For embeddings (optional, for better vectorization)
sentence-transformers>=2.2.0

# For LLM integration (optional)
# anthropic>=0.7.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
ruff>=0.0.290
